backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Subrata Mondal <subratasubha2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pydantic = "^2.6.1"
black = "^24.2.0"
pytest = "^8.0.1"
pytest-cov = "^4.1.0"
pylint = "^3.0.3"
passlib = "^1.7.4"
sqlalchemy = "^2.0.27"
pyjwt = "^2.8.0"
python-decouple = "^3.8"
pydantic-settings = "^2.2.1"
email-validator = "^2.1.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
